Rename VendorAPI::OpenAPISpec
This class is used in the APIDocs namespace as well as the VendorAPI,
which means it should live at the top-level.
