+++ app/controllers/api_docs/openapi_controller.rb
@@ -1,7 +1,7 @@
 module APIDocs
   class OpenapiController < APIDocsController
     def spec
-      render plain: VendorAPI::OpenAPISpec.as_yaml, content_type: 'text/yaml'
+      render plain: VendorAPISpecification.as_yaml, content_type: 'text/yaml'
     end
   end
 end
