+++ spec/lib/openapi_spec.rb
@@ -1,7 +1,7 @@
 require 'rails_helper'
 
 RSpec.describe 'OpenAPI spec' do
-  document = Openapi3Parser.load(VendorAPI::OpenAPISpec.as_hash)
+  document = Openapi3Parser.load(VendorAPISpecification.as_hash)
 
   it 'is a valid OpenAPI spec' do
     expect(document).to be_valid, document.errors.to_a.inspect
