+++ spec/presenters/api_docs/api_schema_spec.rb
@@ -3,7 +3,7 @@ require 'rails_helper'
 RSpec.describe APIDocs::APISchema do
   describe '#object_schema_name' do
     let :schema do
-      @document = Openapi3Parser.load(VendorAPI::OpenAPISpec.as_hash)
+      @document = Openapi3Parser.load(VendorAPISpecification.as_hash)
       schema_name, raw_schema = @document.components.schemas.find { |schema_name, _schema| schema_name == 'ApplicationAttributes' }
       APIDocs::APISchema.new(name: schema_name, schema: raw_schema)
     end
